requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 25.1.1
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
accelerate=1.5.2=pypi_0
aiohappyeyeballs=2.6.1=pyhd8ed1ab_0
aiohttp=3.11.13=pyhe1237c8_0
aiosignal=1.3.2=pyhd8ed1ab_0
annotated-types=0.7.0=pypi_0
arrow-cpp=19.0.0=h865e1df_0
async-timeout=5.0.1=pyhd8ed1ab_1
attrs=25.3.0=pyh71513ae_0
audioread=3.0.1=pypi_0
aws-c-auth=0.6.19=h5eee18b_0
aws-c-cal=0.5.20=hdbd6064_0
aws-c-common=0.8.5=h5eee18b_0
aws-c-compression=0.2.16=h5eee18b_0
aws-c-event-stream=0.2.15=h6a678d5_0
aws-c-http=0.6.25=h5eee18b_0
aws-c-io=0.13.10=h5eee18b_0
aws-c-mqtt=0.7.13=h5eee18b_0
aws-c-s3=0.1.51=hdbd6064_0
aws-c-sdkutils=0.1.6=h5eee18b_0
aws-checksums=0.1.13=h5eee18b_0
aws-crt-cpp=0.18.16=h6a678d5_0
aws-sdk-cpp=1.11.212=hecad206_0
blas=1.0=mkl
boost-cpp=1.78.0=he72f1d9_0
bottleneck=1.4.2=py311hf4808d0_0
brotli-python=1.0.9=py311h6a678d5_9
bzip2=1.0.8=h5eee18b_6
c-ares=1.19.1=h5eee18b_0
ca-certificates=2025.2.25=h06a4308_0
certifi=2025.1.31=pyhd8ed1ab_0
cffi=1.17.1=py311h1fdaa30_1
charset-normalizer=3.4.1=pyhd8ed1ab_0
click=8.1.8=pypi_0
colorama=0.4.6=pyhd8ed1ab_1
contourpy=1.3.1=pypi_0
cpython=3.11.11=py311hd8ed1ab_2
cuda-cudart=12.4.127=0
cuda-cupti=12.4.127=0
cuda-libraries=12.4.1=0
cuda-nvrtc=12.4.127=0
cuda-nvtx=12.4.127=0
cuda-opencl=12.8.90=0
cuda-runtime=12.4.1=0
cuda-version=12.8=3
cycler=0.12.1=pypi_0
datasets=3.4.0=pyhd8ed1ab_0
decorator=5.2.1=pypi_0
dill=0.3.8=pyhd8ed1ab_0
docker-pycreds=0.4.0=pypi_0
filelock=3.18.0=pyhd8ed1ab_0
fonttools=4.56.0=pypi_0
frozenlist=1.5.0=pyhe1237c8_1
fsspec=2024.12.0=pyhd8ed1ab_0
gflags=2.2.2=he1b5a44_1004
gitdb=4.0.12=pypi_0
gitpython=3.1.44=pypi_0
glog=0.5.0=h48cff8f_0
gmp=6.3.0=h6a678d5_0
gmpy2=2.2.1=py311h5eee18b_0
h2=4.2.0=pyhd8ed1ab_0
hpack=4.1.0=pyhd8ed1ab_0
huggingface_hub=0.29.3=pyhd8ed1ab_0
hyperframe=6.1.0=pyhd8ed1ab_0
icu=70.1=h27087fc_0
idna=3.10=pyhd8ed1ab_1
intel-openmp=2023.1.0=hdb19cb5_46306
jinja2=3.1.6=pyhd8ed1ab_0
joblib=1.4.2=pyhd8ed1ab_1
kiwisolver=1.4.8=pypi_0
krb5=1.20.1=h143b758_1
lame=3.100=h7f98852_1001
lazy-loader=0.4=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libabseil=20240116.2=cxx17_h6a678d5_0
libbrotlicommon=1.0.9=h166bdaf_7
libbrotlidec=1.0.9=h166bdaf_7
libbrotlienc=1.0.9=h166bdaf_7
libcublas=12.4.5.8=0
libcufft=11.2.1.3=0
libcufile=1.13.1.3=0
libcurand=10.3.9.90=0
libcurl=8.12.1=hc9e6f67_0
libcusolver=11.6.1.9=0
libcusparse=12.3.1.170=0
libedit=3.1.20230828=h5eee18b_0
libev=4.33=h516909a_1
libevent=2.1.12=hdbd6064_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=13.2.0=h69a702a_0
libgfortran5=13.2.0=ha4646dd_0
libgomp=11.2.0=h1234567_1
libgrpc=1.62.2=h2d74bed_0
libnghttp2=1.57.0=h2d74bed_0
libnpp=12.2.5.30=0
libnvfatbin=12.8.90=0
libnvjitlink=12.4.127=0
libnvjpeg=12.3.1.117=0
libpng=1.6.37=h21135ba_2
libprotobuf=4.25.3=he621ea3_0
librosa=0.11.0=pypi_0
libssh2=1.11.1=h251f7ec_0
libstdcxx-ng=11.2.0=h1234567_1
libthrift=0.15.0=h1795dd8_2
libuuid=1.41.5=h5eee18b_0
llvm-openmp=12.0.1=h4bd325d_1
llvmlite=0.44.0=pypi_0
lz4-c=1.9.4=h6a678d5_1
mad=0.15.1b=h9c3ff4c_1
markupsafe=3.0.2=pyhe1237c8_1
matplotlib=3.10.1=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py311h5eee18b_2
mkl_fft=1.3.11=py311h5eee18b_0
mkl_random=1.2.8=py311ha02d727_0
mpc=1.3.1=h5eee18b_0
mpfr=4.2.1=h5eee18b_0
mpmath=1.3.0=pyhd8ed1ab_1
msgpack=1.1.0=pypi_0
multidict=6.1.0=pyh62beb40_2
multiprocess=0.70.15=py311h06a4308_0
narwhals=1.30.0=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.4.2=pyh267e887_2
numba=0.61.0=pypi_0
numexpr=2.10.1=py311h3c60e43_0
numpy=2.0.1=py311h08b1b3b_1
numpy-base=2.0.1=py311hf175353_1
ocl-icd=2.3.2=h5eee18b_1
openssl=3.0.16=h5eee18b_0
orc=2.0.1=h2d29ad5_0
packaging=24.2=pyhd8ed1ab_2
pandas=2.2.3=py311h6a678d5_0
pillow=11.1.0=pypi_0
pip=25.0=py311h06a4308_0
platformdirs=4.3.6=pypi_0
plotly=6.0.0=pypi_0
pooch=1.8.2=pypi_0
propcache=0.2.1=pyhe1237c8_1
protobuf=5.29.3=pypi_0
psutil=7.0.0=pypi_0
pyarrow=19.0.0=py311h6a678d5_1
pycparser=2.22=pyh29332c3_1
pydantic=2.10.6=pypi_0
pydantic-core=2.27.2=pypi_0
pyparsing=3.2.1=pypi_0
pysocks=1.7.1=pyha55dd90_7
python=3.11.11=he870216_0
python-dateutil=2.9.0.post0=pyhff2d567_1
python-tzdata=2025.1=pyhd8ed1ab_0
python-xxhash=3.5.0=py311h5eee18b_0
python_abi=3.11=2_cp311
pytorch=2.5.1=py3.11_cuda12.4_cudnn9.1.0_0
pytorch-cuda=12.4=hc786d27_7
pytorch-mutex=1.0=cuda
pytz=2025.1=pyhd8ed1ab_0
pyyaml=6.0.2=pyhe1237c8_2
re2=2022.04.01=h27087fc_0
readline=8.2=h5eee18b_0
regex=2024.11.6=py311h5eee18b_0
requests=2.32.3=pyhd8ed1ab_1
s2n=1.3.27=hdbd6064_0
safetensors=0.5.3=py311h448de89_0
scikit-learn=1.6.1=py311h6a678d5_0
scipy=1.15.1=py311h08b1b3b_0
seaborn=0.13.2=pypi_0
sentry-sdk=2.22.0=pypi_0
setproctitle=1.3.5=pypi_0
setuptools=75.8.0=py311h06a4308_0
six=1.17.0=pyhd8ed1ab_0
smmap=5.0.2=pypi_0
snappy=1.2.1=h6a678d5_0
soundfile=0.13.1=pypi_0
sox=14.4.2=h27f72bc_1013
soxr=0.5.0.post1=pypi_0
sqlite=3.45.3=h5eee18b_0
sympy=1.13.1=pypi_0
tbb=2021.5.0=h924138e_1
threadpoolctl=3.6.0=pyhecae5ae_0
tk=8.6.14=h39e8969_0
tokenizers=0.21.0=py311h5e3497e_0
torchaudio=2.5.1=py311_cu124
torchtriton=3.1.0=py311
tqdm=4.67.1=pyhd8ed1ab_1
transformers=4.50.2=pypi_0
typing-extensions=4.12.2=hd8ed1ab_1
typing_extensions=4.12.2=pyha770c72_1
tzdata=2025a=h04d1e81_0
urllib3=2.3.0=pyhd8ed1ab_0
utf8proc=2.6.1=h5eee18b_1
wandb=0.19.8=pypi_0
wheel=0.45.1=py311h06a4308_0
xxhash=0.8.0=h7f98852_3
xz=5.6.4=h5eee18b_1
yaml=0.2.5=h7f98852_2
yarl=1.18.3=pyhe1237c8_1
zlib=1.2.13=h5eee18b_1
zstandard=0.23.0=py311h2c38b39_1
zstd=1.5.6=hc292b87_0
